refactor: extract decompressBuffer to shared util (src/utils/decompress.ts)
Eliminates duplicate decompression logic between global-monitor.ts and
RequestMonitoringService.ts. The shared util uses string | undefined for
the encoding parameter — IncomingMessage.headers['content-encoding'] is
always a string at runtime, so the array branch in the old service copy
was dead code.
