+++ src/global-monitor.ts
@@ -9,6 +9,7 @@ import { CoolhandCallData, CoolhandRequestOptions, CoolhandMatchedPattern } from
 import { PatternMatchingService } from './services/PatternMatchingService.js';
 import { LoggingService } from './services/LoggingService.js';
 import { parseBody } from './utils/parse-body.js';
+import { decompressBuffer } from './utils/decompress.js';
 
 type HttpClientRequest = any; // Will be properly typed when http is loaded
 type HttpIncomingMessage = any; // Will be properly typed when http is loaded
@@ -23,7 +24,6 @@ const isEdgeRuntime = () => {
 // Node.js modules - conditionally imported
 let https: any = null;
 let http: any = null;
-let zlib: any = null;
 
 // Lazy load Node.js modules only when not in Edge runtime
 const loadNodeModules = async () => {
@@ -54,13 +54,6 @@ const loadNodeModules = async () => {
 
     https = httpsModule;
     http = httpModule;
-
-    try {
-      const zlibModule = await import('zlib');
-      zlib = zlibModule;
-    } catch {
-      log('zlib not available — response decompression disabled');
-    }
 
     return true;
   } catch {
@@ -392,56 +385,6 @@ function patchFetch(): void {
   }
 }
 
-function decompressBuffer(buffer: Buffer, encoding: string | undefined): Promise<string> {
-  return new Promise((resolve) => {
-    if (!encoding || !zlib) {
-      resolve(buffer.toString('utf-8'));
-      return;
-    }
-
-    const enc = encoding.trim().toLowerCase();
-
-    if (enc === 'gzip' || enc === 'x-gzip') {
-      zlib.gunzip(buffer, (err: Error | null, result: Buffer) => {
-        if (err) {
-          log(`⚠️ Decompression failed for encoding '${enc}': ${err.message}`);
-          resolve(buffer.toString('utf-8'));
-        } else {
-          resolve(result.toString('utf-8'));
-        }
-      });
-    } else if (enc === 'deflate') {
-      // RFC 1950 (zlib-wrapped) first; fall back to RFC 1951 (raw deflate)
-      // as some servers (older IIS, some load balancers) send raw deflate despite
-      // the content-encoding header implying the zlib wrapper.
-      zlib.inflate(buffer, (err: Error | null, result: Buffer) => {
-        if (!err) {
-          resolve(result.toString('utf-8'));
-          return;
-        }
-        zlib.inflateRaw(buffer, (rawErr: Error | null, rawResult: Buffer) => {
-          if (rawErr) {
-            log(`⚠️ Decompression failed for encoding 'deflate': ${rawErr.message}`);
-            resolve(buffer.toString('utf-8'));
-          } else {
-            resolve(rawResult.toString('utf-8'));
-          }
-        });
-      });
-    } else if (enc === 'br') {
-      zlib.brotliDecompress(buffer, (err: Error | null, result: Buffer) => {
-        if (err) {
-          log(`⚠️ Decompression failed for encoding '${enc}': ${err.message}`);
-          resolve(buffer.toString('utf-8'));
-        } else {
-          resolve(result.toString('utf-8'));
-        }
-      });
-    } else {
-      resolve(buffer.toString('utf-8'));
-    }
-  });
-}
 
 function interceptRequest(
   originalRequest: any,
@@ -490,7 +433,7 @@ function interceptRequest(
       try {
         const rawBuffer = Buffer.concat(responseChunks);
         const contentEncoding = res.headers?.['content-encoding'];
-        const responseBody = await decompressBuffer(rawBuffer, contentEncoding);
+        const responseBody = await decompressBuffer(rawBuffer, contentEncoding, log);
         callData.response_body = parseBody(responseBody);
       } catch (err: any) {
         log(`⚠️ Response body capture failed for call #${callData.id}: ${err?.message}`);
