+++ src/services/RequestMonitoringService.ts
@@ -1,9 +1,9 @@
 import * as https from 'https';
 import * as http from 'http';
-import * as zlib from 'zlib';
 import { CoolhandCallData, CoolhandRequestOptions, CoolhandMatchedPattern } from '../types';
 import { PatternMatchingService } from './PatternMatchingService.js';
 import { parseBody } from '../utils/parse-body.js';
+import { decompressBuffer } from '../utils/decompress.js';
 
 export class RequestMonitoringService {
   private callCounter: number = 0;
@@ -214,7 +214,7 @@ export class RequestMonitoringService {
         try {
           const rawBuffer = Buffer.concat(responseChunks);
           const contentEncoding = res.headers?.['content-encoding'];
-          const responseBody = await this.decompressBuffer(rawBuffer, contentEncoding);
+          const responseBody = await decompressBuffer(rawBuffer, contentEncoding, this.log.bind(this));
           callData.response_body = parseBody(responseBody);
         } catch (err: any) {
           this.log(`⚠️ Response body capture failed for call #${callData.id}: ${err?.message}`);
@@ -303,58 +303,6 @@ export class RequestMonitoringService {
       this.log(`❌ Fetch error for call #${callData.id}:`, (error as Error).message);
       throw error;
     }
-  }
-
-  private decompressBuffer(buffer: Buffer, encoding: string | string[] | undefined): Promise<string> {
-    return new Promise((resolve) => {
-      const rawEncoding = Array.isArray(encoding) ? encoding[0] : encoding;
-      if (!rawEncoding) {
-        resolve(buffer.toString('utf-8'));
-        return;
-      }
-
-      const enc = rawEncoding.trim().toLowerCase();
-
-      if (enc === 'gzip' || enc === 'x-gzip') {
-        zlib.gunzip(buffer, (err, result) => {
-          if (err) {
-            this.log(`⚠️ Decompression failed for encoding '${enc}': ${err.message}`);
-            resolve(buffer.toString('utf-8'));
-          } else {
-            resolve(result.toString('utf-8'));
-          }
-        });
-      } else if (enc === 'deflate') {
-        // RFC 1950 (zlib-wrapped) first; fall back to RFC 1951 (raw deflate)
-        // as some servers (older IIS, some load balancers) send raw deflate despite
-        // the content-encoding header implying the zlib wrapper.
-        zlib.inflate(buffer, (err, result) => {
-          if (!err) {
-            resolve(result.toString('utf-8'));
-            return;
-          }
-          zlib.inflateRaw(buffer, (rawErr, rawResult) => {
-            if (rawErr) {
-              this.log(`⚠️ Decompression failed for encoding 'deflate': ${rawErr.message}`);
-              resolve(buffer.toString('utf-8'));
-            } else {
-              resolve(rawResult.toString('utf-8'));
-            }
-          });
-        });
-      } else if (enc === 'br') {
-        zlib.brotliDecompress(buffer, (err, result) => {
-          if (err) {
-            this.log(`⚠️ Decompression failed for encoding '${enc}': ${err.message}`);
-            resolve(buffer.toString('utf-8'));
-          } else {
-            resolve(result.toString('utf-8'));
-          }
-        });
-      } else {
-        resolve(buffer.toString('utf-8'));
-      }
-    });
   }
 
   private buildURL(options: CoolhandRequestOptions | string | URL, protocol: string): string {
